+++ ethereum/main.go
@@ -62,5 +62,5 @@ func main() {
 		panic(err)
 	}
 
-	c.SendToHolding(context.TODO(), 42)
+	c.SendToHolding(context.TODO(), 1)
 }